% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectErythropoiesis.R
\docType{class}
\name{TrenaProjectErythropoiesis-class}
\alias{TrenaProjectErythropoiesis-class}
\alias{.TrenaProjectErythropoiesis}
\alias{TrenaProjectErythropoiesis}
\title{TrenaProjectErythropoiesis-class}
\usage{
TrenaProjectErythropoiesis(quiet = TRUE)
}
\value{
An object of the TrenaProjectErythropoiesis class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
